Fix path of config file on unix systems (linux/osx)
Without this fix, on filesystems that don't use '\' as a path 
separator, the config gets put in a file named: 
"config\Camping\The Camping Mod.cfg" with the '\'s as part of the
filename. Which is not ideal. I've tested this fix on windows, 
linux and osx, but it would probably be better to use something like 
this: http://stackoverflow.com/questions/711993/does-java-have-a-path-joining-method
to join the paths properly.